Create DelegateCall.sol
Fortunately, there is the EIP-1167 specification that defines how to implement a proxy contract cheaply. This proxy will forward all calls and 100% of the gas to the implementation contract and then relay the return value back to the caller. 

The whole magic of the proxy contract is done using delegatecall. You can get an idea of how it works by reading this article.
Let’s see now how to make this work. First, you need to grab the implementation of this specification, which you can find on GitHub. Copy-paste the code of CloneFactory in your project.